\name{capply}
\alias{capply}
\title{Character apply}
\usage{
  capply(objects, FUN, simplify = TRUE, ...)
}
\arguments{
  \item{objects}{Character vector with object names}

  \item{FUN}{Function to apply}

  \item{...}{Additional arguments to pass to FUN}
}
\description{
  Pass several objects through a function where the objects
  are refered to by character strings.
}

