\name{merge.formula}
\alias{merge.formula}
\title{Merge two formulas}
\usage{
  merge.formula(form1, form2, func = union, ...)
}
\arguments{
  \item{form1}{Formula object.}

  \item{form2}{Formula object.}

  \item{func}{Function for merging the terms in
  \code{form1} and \code{form2}.}

  \item{...}{Addition params.}
}
\description{
  Merge two formulas
}
\references{
  Formula object.
}

