\docType{package}
\name{pr2-package}
\alias{pr2}
\alias{pr2-package}
\title{The pr2 Package}
\description{
  Predictive R-squared.
}
\details{
  \tabular{ll}{ Package: \tab pr2\cr Type: \tab Package\cr
  Version: \tab 0.1\cr Date: \tab 2013-03-05\cr Depends:
  \tab abind, vegan\cr License: \tab GPL (>= 3)\cr
  LazyLoad: \tab yes\cr URL: \tab
  http://stevencarlislewalker.wordpress.com/\cr }

  Tools for predictive R-square analysis
}
\author{
  Steve Walker
}
\keyword{package}

