\name{rep.Rsqr.sims}
\alias{rep.Rsqr.sims}
\title{Replicate community simlations for testing R squares}
\usage{
  rep.Rsqr.sims(iters, ...)
}
\arguments{
  \item{iters}{Number of replicate communities}

  \item{...}{Additional arguments to pass to
  \code{\link{Rsqr.sims}}}
}
\value{
  List (or is it a data frame) of class \code{Rsqr.sims} of
  replicate results.
}
\description{
  Replicate community simlations for testing R squares
}
\seealso{
  Other sims: \code{\link{Rsqr.sims}}
}

